requirements.txt
scrapy
scrapy-djangoitem
Django
dateparser
geocoder
psycopg2
djangorestframework
django-filter
whitenoise
brotlipy